requirements.txt
# Python dependencies for Drift Phonk Autopilot Playbook

# Core dependencies
pyyaml>=6.0
numpy>=1.21.0
scipy>=1.9.0

# Audio processing
librosa>=0.9.0
sounddevice>=0.4.4
pyaudio>=0.2.11

# Real-time processing
threading>=3.8
asyncio>=3.8

# Data handling
pandas>=1.3.0

# Optional: For Tesla autopilot integration (requires additional setup)
# teslapy>=2.0.0

# Optional: For Comma AI integration (requires additional setup)  
# openpilot>=0.8.0

# Development dependencies (install with: pip install -r requirements.txt -r requirements-dev.txt)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
# mypy>=0.950